Minor refactorings, rename tab-created to tab-updated, move tests around to separate unit and e2e tests. Add tests to verify that new deduplication behavior works as expected.
