+++ tests/unit/handlers/tab-focused.test.js
