+++ tests/unit/handlers/tab-updated.test.js
