+++ tests/unit/tabs/tab-bookmarker.test.js
@@ -1,4 +1,4 @@
-import TabBookmarker from '../../src/tab/tab-bookmarker';
+import TabBookmarker from '../../../src/tab/tab-bookmarker';
 import { assert } from 'chai';
 
 const chrome = require('sinon-chrome/extensions');
