+++ tests/unit/tabs/tab-deduplicator.test.js
@@ -1,7 +1,10 @@
-import TabDeduplicator from '../../src/tab/tab-deduplicator';
+import TabDeduplicator from '../../../src/tab/tab-deduplicator';
 import { assert } from 'chai';
+import sinon from 'sinon/pkg/sinon-esm';
 
+// TODO: investigate https://github.com/acvetkov/sinon-chrome/pull/94 and fork to mock missing APIs
 const chrome = require('sinon-chrome/extensions');
+
 
 describe('tab-deduplicator', () => {
 	let tabDeduplicator;
@@ -9,6 +12,7 @@ describe('tab-deduplicator', () => {
 
 	beforeEach(() => {
 		global.chrome = chrome;
+		chrome.tabs.goBack = sinon.stub();
 		tabLock = new Set();
 		tabDeduplicator = new TabDeduplicator(tabLock);
 	});
@@ -36,23 +40,36 @@ describe('tab-deduplicator', () => {
 		assert(chrome.tabs.remove.notCalled);
 	});
 
-	it('should deduplicate tab', async () => {
+	it('should remove tab if tab has openerTabId', async () => {
+		await tabDeduplicator.deduplicateTab(
+			{ id: 1, url: 'theo.lol', status: 'loading', active: true, openerTabId: 2 },
+			[{ id: 2, url: 'theo.lol' }],
+		);
+		assert(chrome.tabs.highlight.calledOnce);
+		assert(chrome.windows.update.calledOnce);
+		assert(chrome.tabs.remove.calledOnce);
+		assert(chrome.tabs.goBack.notCalled);
+	});
+
+	it('should call goBack on tab if tab doesnt have openerTabId', async () => {
 		await tabDeduplicator.deduplicateTab(
 			{ id: 1, url: 'theo.lol', status: 'loading', active: true },
 			[{ id: 2, url: 'theo.lol' }],
 		);
+		assert(chrome.tabs.goBack.calledOnce);
 		assert(chrome.tabs.highlight.calledOnce);
 		assert(chrome.windows.update.calledOnce);
-		assert(chrome.tabs.remove.calledOnce);
+		assert(chrome.tabs.remove.notCalled);
 	});
 
 	it('shouldnt focus tab if not active', async () => {
 		await tabDeduplicator.deduplicateTab(
-			{ id: 1, url: 'theo.lol', status: 'loading', active: false },
+			{ id: 1, url: 'theo.lol', status: 'loading', active: false, openerTabId: 2 },
 			[{ id: 2, url: 'theo.lol' }],
 		);
 		assert(chrome.tabs.highlight.calledOnce);
 		assert(chrome.windows.update.calledOnce);
 		assert(chrome.tabs.remove.calledOnce);
+		assert(chrome.tabs.goBack.notCalled);
 	});
 });
