+++ tests/unit/tabs/tab-grouper.test.js
@@ -1,5 +1,5 @@
-import TabGrouper from '../../src/tab/tab-grouper';
-import sinon from '../../node_modules/sinon/pkg/sinon-esm.js';
+import TabGrouper from '../../../src/tab/tab-grouper';
+import sinon from '../../../node_modules/sinon/pkg/sinon-esm.js';
 import { assert } from 'chai';
 
 const chrome = require('sinon-chrome/extensions');
