+++ tests/unit/tabs/tab-pruner.test.js
@@ -1,5 +1,5 @@
-import TabBookmarker from '../../src/tab/tab-bookmarker';
-import TabPruner from '../../src/tab/tab-pruner';
+import TabBookmarker from '../../../src/tab/tab-bookmarker';
+import TabPruner from '../../../src/tab/tab-pruner';
 import { assert } from 'chai';
 
 const chrome = require('sinon-chrome/extensions');
