+++ tests/unit/tabs/tab-tracker.test.ts
@@ -1,9 +1,9 @@
-import TabTracker from '../../src/tab/tab-tracker';
+import TabTracker from '../../../src/tab/tab-tracker';
 import { assert } from 'chai';
 
 import chrome from 'sinon-chrome/extensions';
 import sinon from 'sinon';
-import { PruneStorage } from '../../src/util/storage';
+import { PruneStorage } from '../../../src/util/storage';
 declare var global: any;
 
 describe('tab-tracker', () => {
